refactor(static-verifier): move keygen and EVM code to dedicated keygen module
Move StaticVerifierProvingKey, Halo2Prover::keygen, and all EVM types
and functions out of prover/ into a top-level keygen.rs module. The
prover module now only contains proving/verification logic and shared
types.
