+++ crates/static-verifier/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod config;
 pub mod field;
 pub mod hash;
+pub mod keygen;
 pub mod prover;
 pub mod stages;
 pub mod transcript;
@@ -13,8 +14,9 @@ pub use config::{
     STATIC_VERIFIER_NUM_ADVICE_COLS_PHASE0,
 };
 pub use halo2_base::halo2_proofs::halo2curves::bn256::Fr;
+pub use keygen::StaticVerifierProvingKey;
 pub use openvm_stark_sdk::config::baby_bear_bn254_poseidon2::{EF as ChildEF, F as ChildF};
 pub use prover::{
     Halo2Params, Halo2Prover, Halo2ProvingMetadata, Halo2ProvingPinning, StaticVerifierInput,
-    StaticVerifierProof, StaticVerifierProvingKey,
+    StaticVerifierProof,
 };
